chore : add category form functionality
