+++ package.json
@@ -38,6 +38,7 @@
     "@react-email/components": "^0.0.25",
     "@tanstack/react-table": "^8.20.5",
     "@uploadthing/react": "^7.0.2",
+    "axios": "^1.7.7",
     "bcryptjs": "^2.4.3",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
