--- src/app/(pages)/dashboard/category/[categoryId]/page.tsx
@@ -1,28 +1,0 @@
-import { Breadcrumbs } from '@/components/breadcrumbs';
-import { ProductForm } from '@/components/forms/product-form';
-import PageContainer from '@/components/layout/page-container';
-// import { ScrollArea } from '@/components/ui/scroll-area';
-import React from 'react';
-
-const breadcrumbItems = [
-  { title: 'Dashboard', link: '/dashboard' },
-  { title: 'Categories', link: '/dashboard/categories' },
-  { title: 'Create', link: '/dashboard/categories/create' }
-];
-export default function Page() {
-  return (
-    <PageContainer scrollable={true}>
-      <div className="space-y-4">
-        <Breadcrumbs items={breadcrumbItems} />
-        <ProductForm
-          categories={[
-            { _id: 'shirts', name: 'shirts' },
-            { _id: 'pants', name: 'pants' }
-          ]}
-          initialData={null}
-          key={null}
-        />
-      </div>
-    </PageContainer>
-  );
-}
